+++ src/components/layouts/Features/index.tsx
@@ -13,14 +13,10 @@ import {
   VStack,
   useMediaQuery,
 } from '@chakra-ui/react';
-import { linedcUrl } from '../../../config/constants';
+import { globalDescription, linedcUrl } from '../../../config/constants';
 import { OwnSNS } from '../OwnSNS';
 
-const description = `
-1991年1月生まれ福岡出身。九州大学理学部物理学科卒業後、都内SIerにてインフラエンジニアとして働いた後、ミロゴス株式会社でフロント・バックエンド・インフラをフルスタックにこなすエンジニアを経験。
-現在はCyberAgentのAI事業本部でバックエンドエンジニアとして従事中。
-好きな技術はFlutter/Next.js/AWS。現在娘の子育てに奮闘中。
-`;
+const description = globalDescription;
 
 export default function Features() {
   const [isLagerThan872] = useMediaQuery('(min-width: 872px)');
