+++ src/pages/articles/index.tsx
@@ -1,3 +1,5 @@
+import Seo, { MetaData } from '@/components/layouts/Seo';
+import { baseURL } from '@/config/constants';
 import { GetStaticProps } from 'next';
 import { ArticleList } from '../../features/article/ArticleList.page';
 import { fetchArticles } from '../../features/article/apis/article';
@@ -8,8 +10,14 @@ interface ArticlePageProps {
 }
 
 export default function ArticlesPage({ articles }: ArticlePageProps) {
+  const metaData: MetaData = {
+    pageTitle: '記事一覧',
+    pagePath: `${baseURL}/news`,
+  };
+
   return (
     <>
+      <Seo {...metaData} />
       <ArticleList articles={articles} />
     </>
   );
