+++ src/pages/news/[news_id]/index.tsx
@@ -1,3 +1,5 @@
+import Seo, { MetaData } from '@/components/layouts/Seo';
+import { baseURL } from '@/config/constants';
 import { GetStaticPaths, GetStaticProps } from 'next';
 import { ParsedUrlQuery } from 'querystring';
 import { NewsDetail } from '../../../features/news/NewsDetail.page';
@@ -9,8 +11,16 @@ interface NewsDetailPageProps {
 }
 
 export default function NewsDetailPage({ news }: NewsDetailPageProps) {
+  const metaData: MetaData = {
+    pageTitle: news.title,
+    pageDescription: news.body,
+    pagePath: `${baseURL}/news/${news.id}`,
+    pageImg: news.image.url,
+  };
+
   return (
     <>
+      <Seo {...metaData} />
       <NewsDetail news={news} />
     </>
   );
